feat: remove planning warning dialog and adapt texts
